requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==2.6.0
transformers==4.36.2
faiss-cpu==1.7.4
Pillow==10.2.0
requests==2.31.0
numpy==1.24.3
pillow-avif-plugin==1.2.0
tqdm==4.66.1